a little house keeping with removing old files. also did another test run and saw that crop_random won't work since our images can't change shape in augmentation
