+++ codes/block_definitions/utilities/operators_TensorFlow_Keras.py
@@ -43,7 +43,6 @@ def conv2D_layer(input_tensor, filters=64, kernel_size=3, activation=tf.nn.relu)
                                   data_format="channels_last"
                                  )(input_tensor)
 
-
 operator_dict[conv2D_layer] = {"inputs": [tf.keras.layers],
                                "output": tf.keras.layers,
                                "args": [argument_types.ArgumentType_Pow2,
@@ -63,7 +62,6 @@ def conv2DTranspose_layer(input_tensor, filters=64, kernel_size=3, activation=tf
                                            activation=activation,
                                            data_format="channels_last"
                                            )(input_tensor)
-
 
 operator_dict[conv2DTranspose_layer] = {"inputs": [tf.keras.layers],
                                         "output": tf.keras.layers,
@@ -117,8 +115,73 @@ def gaussianNoise_layer(input_tensor, stddev):
     '''
     return tf.keras.layers.GaussianNoise(stddev=stddev)(input_tensor)
 
-
 operator_dict[gaussianNoise_layer] = {"inputs": [tf.keras.layers],
                                         "output": tf.keras.layers,
                                         "args": [argument_types.ArgumentType_Float0to1]
                                        }
+
+
+### Pooling Layers
+
+def avg_pool_layer(input_tensor, pool_height=2, pool_width=2, strides=2):
+    if input_tensor.rank == 1:
+        return tf.keras.layers.AveragePooling1D(pool_size=[pool_height, pool_width], strides=strides, padding="valid")(input_tensor)
+    if input_tensor.rank == 2:
+        return tf.keras.layers.AveragePooling2D(pool_size=[pool_height, pool_width], strides=strides, padding="valid")(input_tensor)
+    if input_tensor.rank == 3:
+        return tf.keras.layers.AveragePooling3D(pool_size=[pool_height, pool_width], strides=strides, padding="valid")(input_tensor)
+
+operator_dict[avg_pool_layer] = {"inputs": [tf.keras.layers],
+                                 "output": tf.keras.layers,
+                                 "args": [argument_types.ArgumentType_TFPoolSize,
+                                          argument_types.ArgumentType_TFPoolSize,
+                                          argument_types.ArgumentType_TFPoolSize]  # TODO verify the argtype we want for strides...i just picked one
+                                 }
+
+
+def max_pool_layer(input_tensor, pool_height=2, pool_width=2, strides=2):
+    if input_tensor.shape[1].value == 1:
+        return input_tensor
+    else:
+        return tf.keras.layers.MaxPool2D(pool_size=[pool_height, pool_width], strides=strides, padding="valid")(input_tensor)
+
+operator_dict[max_pool_layer] = {"inputs": [tf.keras.layers],
+                                 "output": tf.keras.layers,
+                                 "args": [argument_types.ArgumentType_TFPoolSize,
+                                          argument_types.ArgumentType_TFPoolSize,
+                                          argument_types.ArgumentType_TFPoolSize]
+                                 }
+
+
+def fractional_max_pool(input_tensor, pool_height=2, pool_width=2):
+    if input_tensor.shape[1] == 1:
+        return input_tensor
+    # see args.py for mutation limits
+    pooling_ratio = [1.0, pool_height, pool_width, 1.0]
+    # true random underfits when combined with data augmentation and/or dropout
+    pseudo_random = True
+    # overlapping pooling regions work better, according to 2015 Ben Graham paper
+    overlapping = True
+    # returns a tuple of Tensor objects (output, row_pooling_sequence, col_pooling_sequence
+    return tf.nn.fractional_max_pool(input_tensor, pooling_ratio, pseudo_random, overlapping)[0]
+
+operator_dict[fractional_max_pool] = {"inputs": [tf.keras.layers],
+                                      "args": [argument_types.ArgumentType_TFPoolSize, argument_types.ArgumentType_TFPoolSize],
+                                      "output": tf.keras.layers
+                                      }
+
+
+def fractional_avg_pool(input_tensor, pool_height=2.0, pool_width=2.0):
+    if input_tensor.shape[1].value == 1:
+        return input_tensor
+    pooling_ratio = [1.0, pool_height, pool_width, 1.0]
+    pseudo_random = True
+    overlapping = True
+    # returns a tuple of Tensor objects (output, row_pooling_sequence, col_pooling_sequence)
+    return tf.nn.fractional_avg_pool(input_tensor, pooling_ratio, pseudo_random, overlapping)[0]
+
+operator_dict[fractional_avg_pool] = {"inputs": [tf.keras.layers],
+                                      "args": [argument_types.ArgumentType_TFPoolSize, argument_types.ArgumentType_TFPoolSize],
+                                      "output": tf.keras.layers
+                                      }
+
